// backend/apps/payment-service/src/types/payment.type.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5'
  }
  payments: {
    Tables: {
      payment_methods: {
        Row: {
          created_at: string
          id: string
          method_name: string
        }
        Insert: {
          created_at?: string
          id: string
          method_name: string
        }
        Update: {
          created_at?: string
          id?: string
          method_name?: string
        }
        Relationships: []
      }
      payments: {
        Row: {
          amount: number
          athlete_id: string
          payment_date: string | null
          payment_id: string
          payment_method_id: string
          subscription_id: string | null
          transaction_reference: string | null
        }
        Insert: {
          amount: number
          athlete_id: string
          payment_date?: string | null
          payment_id: string
          payment_method_id: string
          subscription_id?: string | null
          transaction_reference?: string | null
        }
        Update: {
          amount?: number
          athlete_id?: string
          payment_date?: string | null
          payment_id?: string
          payment_method_id?: string
          subscription_id?: string | null
          transaction_reference?: string | null
        }
        Relationships: [
          {
            foreignKeyName: 'payments_payment_method_id_fkey'
            columns: ['payment_method_id']
            isOneToOne: false
            referencedRelation: 'payment_methods'
            referencedColumns: ['id']
          },
          {
            foreignKeyName: 'payments_subscription_id_fkey'
            columns: ['subscription_id']
            isOneToOne: false
            referencedRelation: 'subscriptions'
            referencedColumns: ['id']
          }
        ]
      }
      plans: {
        Row: {
          amount: number
          created_at: string
          currency: string
          description: string | null
          id: string
          interval: 'days' | 'weeks' | 'months' | 'years'
          interval_count: number
          is_active: boolean
          name: string
          updated_at: string
        }
        Insert: {
          amount: number
          created_at?: string
          currency?: string
          description?: string | null
          id?: string
          interval: 'days' | 'weeks' | 'months' | 'years'
          interval_count?: number
          is_active?: boolean
          name: string
          updated_at?: string
        }
        Update: {
          amount?: number
          created_at?: string
          currency?: string
          description?: string | null
          id?: string
          interval?: 'days' | 'weeks' | 'months' | 'years'
          interval_count?: number
          is_active?: boolean
          name?: string
          updated_at?: string
        }
        Relationships: []
      }
      subscriptions: {
        Row: {
          athlete_id: string
          created_at: string | null
          end_date: string
          id: string
          start_date: string
          status: string
          updated_at: string | null
        }
        Insert: {
          athlete_id: string
          created_at?: string | null
          end_date: string
          id?: string
          start_date: string
          status?: string
          updated_at?: string | null
        }
        Update: {
          athlete_id?: string
          created_at?: string | null
          end_date?: string
          id?: string
          start_date?: string
          status?: string
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema['CompositeTypes']
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes']
    : never = never
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions['schema']]['CompositeTypes'][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema['CompositeTypes']
    ? DefaultSchema['CompositeTypes'][PublicCompositeTypeNameOrOptions]
    : never